Truncations

Currently, truncations are supported through the following different methods:

MatrixAlgebraKit.truncrank — Function

truncrank(howmany::Int, by=abs, rev=true)

Truncation strategy to keep the first howmany values when sorted according to by or the last howmany if rev is true.

MatrixAlgebraKit.trunctol — Function

trunctol(atol::Real)

Truncation strategy to discard the values that are smaller than atol in absolute value.

MatrixAlgebraKit.truncabove — Function

truncabove(atol::Real)

Truncation strategy to discard the values that are larger than atol in absolute value.